Bundle Lambda functions using esbuild
esbuild is installed as a non-dev dependency because SAM expects esbuild to be available on the PATH or as a non-dev dependency.

Bundling as esm doesn't work because a dependency has a dynamic import, which esbuild doesn't support for esm.

Bundling for cjs works but doesn't support top-level await, so RP_ID_HASH is now awaited at time of use.