Use separate user IDs and handles
This avoids an API call and potential TOCTOU issues with how IDs are picked, though does mean an extra 64 bytes are stored per user (plus 64 bytes per passkey to avoid an additional DB lookup).

Also don't send the user handle to the backend when adding an additional passkey as it's already known, and ensure that the user handle is present when signing in.